Convert EventSettings, Match, and MatchResult models to use Bolt DB.

+++ field/arena_notifiers.go
@@ -85,7 +85,7 @@ func (arena *Arena) generateArenaStatusMessage() interface{} {
 	}
 
 	return &struct {
-		MatchId          int
+		MatchId          int64
 		AllianceStations map[string]*AllianceStation
 		TeamWifiStatuses map[string]network.TeamWifiStatus
 		MatchState
